+++ src/Cronus.Serialization.NewtonsoftJson.Benchmarks/IListData.cs
@@ -1,0 +1,4 @@
+﻿public interface IListData
+{
+    List<Data> TheData { get; }
+}