+++ src/Elders.Cronus.Serialization.NewtonsoftJson.Models/SimpleType.cs
@@ -1,8 +1,45 @@
 using System;
+using System.Collections.Generic;
 using System.Runtime.Serialization;
 
 namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests
 {
+    [DataContract(Name = "TypeWithCollection")]
+    public class TypeWithCollection
+    {
+        public TypeWithCollection()
+        {
+            Collection = new List<TypeWithCollectionItem>();
+        }
+
+        public TypeWithCollection(int id) : this()
+        {
+            Id = id;
+        }
+
+        [DataMember(Order = 1)]
+        public int Id { get; set; }
+
+        [DataMember(Order = 2)]
+        public List<TypeWithCollectionItem> Collection { get; set; }
+    }
+
+    [DataContract(Name = "TypeWithCollectionItem")]
+    public class TypeWithCollectionItem
+    {
+        [DataMember(Order = 1)]
+        public string String { get; set; }
+
+        [DataMember(Order = 2)]
+        public int Int { get; set; }
+
+        [DataMember(Order = 3)]
+        public DateTime Date { get; set; }
+
+        [DataMember(Order = 4)]
+        public StructType StructProp { get; set; }
+    }
+
     [DataContract(Name = "SimpleType")]
     public class SimpleType
     {