+++ src/Elders.Cronus.Serialization.NewtonsoftJson.Tests/DomainModeling/When_EntityId__COOOOOL.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Runtime.Serialization;
 using Machine.Specifications;
 
@@ -17,20 +16,16 @@ namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests.custom_cases
             var contracts = new List<Type>();
             contracts.AddRange(typeof(WorkId).Assembly.GetExportedTypes());
             serializer = new JsonSerializer(contracts);
-            serializer2 = new JsonSerializer(contracts);
-            serStream = new MemoryStream();
-            serializer.Serialize(serStream, ser);
-            serStream.Position = 0;
+            data = serializer.SerializeToBytes(ser);
         };
-        Because of_deserialization = () => { deser = (WorkId)serializer2.Deserialize(serStream); };
+        Because of_deserialization = () => deser = serializer.DeserializeFromBytes<WorkId>(data);
 
         It should_not_be_null = () => deser.ShouldNotBeNull();
 
         static WorkId ser;
         static WorkId deser;
-        static Stream serStream;
         static JsonSerializer serializer;
-        static JsonSerializer serializer2;
+        static byte[] data;
     }
 
     [DataContract(Name = "33908fe3-89d8-458f-975f-4a1e273c2134")]