+++ src/Elders.Cronus.Serialization.NewtonsoftJson.Tests/custom-cases/SerializableException.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Runtime.Serialization;
 using Machine.Specifications;
 
@@ -54,6 +53,7 @@ namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests.custom_cases
 
         Establish context = () =>
         {
+            SerializableException ser;
             try
             {
                 try
@@ -73,21 +73,16 @@ namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests.custom_cases
             contracts.AddRange(typeof(NestedType).Assembly.GetExportedTypes());
             contracts.AddRange(typeof(Whem_SerializableException).Assembly.GetExportedTypes());
             serializer = new JsonSerializer(contracts);
-            serializer2 = new JsonSerializer(contracts);
-            serStream = new MemoryStream();
-            serializer.Serialize(serStream, ser);
-            serStream.Position = 0;
+            data = serializer.SerializeToBytes(ser);
         };
 
-        Because of_deserialization = () => { deser = (SerializableException)serializer.Deserialize(serStream); };
+        Because of_deserialization = () => deser = serializer.DeserializeFromBytes<SerializableException>(data);
 
         It should_not_be_null = () => deser.ShouldNotBeNull();
 
-        static SerializableException ser;
         static SerializableException deser;
-        static Stream serStream;
         static JsonSerializer serializer;
-        static JsonSerializer serializer2;
+        static byte[] data;
     }
 
 }