+++ src/Elders.Cronus.Serialization.NewtonsoftJson.Tests/runtime-type-discovery/When_serializng_simple_type_with_headers.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using Machine.Specifications;
 
 namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests.runtime_type_discovery
@@ -14,12 +13,10 @@ namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests.runtime_type_discover
             var contracts = new List<Type>();
             contracts.AddRange(typeof(NestedType).Assembly.GetExportedTypes());
             serializer = new JsonSerializer(contracts);
-            serStream = new MemoryStream();
-            serializer.Serialize(serStream, ser);
-            serStream.Position = 0;
+            data = serializer.SerializeToBytes(ser);
         };
-        Because of_deserialization = () => { deser = (SimpleTypeWithHeaders)serializer.Deserialize(serStream); };
 
+        Because of_deserialization = () => deser = serializer.DeserializeFromBytes<SimpleTypeWithHeaders>(data);
 
         It should_not_be_null = () => deser.ShouldNotBeNull();
         It should_have_the_same_int = () => deser.Int.ShouldEqual(ser.Int);
@@ -29,7 +26,7 @@ namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests.runtime_type_discover
 
         static SimpleTypeWithHeaders ser;
         static SimpleTypeWithHeaders deser;
-        static Stream serStream;
         static JsonSerializer serializer;
+        static byte[] data;
     }
 }