+++ src/Elders.Cronus.Serialization.NewtonsoftJson.Tests/runtime-type-inheritence-model/Whem_serializing__Money__struct.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using Machine.Specifications;
 
 namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests
@@ -15,24 +14,17 @@ namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests
             var contracts = new List<Type>();
             contracts.AddRange(typeof(StructType).Assembly.GetExportedTypes());
             serializer = new JsonSerializer(contracts);
-            serializer2 = new JsonSerializer(contracts);
-            serStream = new MemoryStream();
-            serializer.Serialize(serStream, ser);
-            serStream.Position = 0;
-
-            //var json = Encoding.ASCII.GetString(serStream.ToArray());
-            //serStream.Position = 0;
+            data = serializer.SerializeToBytes(ser);
         };
-        Because of_deserialization = () => deser = (StructType)serializer2.Deserialize(serStream);
 
+        Because of_deserialization = () => deser = serializer.DeserializeFromBytes<StructType>(data);
 
         It should_not_be_null = () => deser.ShouldNotBeNull();
         It should_have_the_same_value = () => deser.ShouldEqual(ser);
 
         static StructType ser;
         static StructType deser;
-        static MemoryStream serStream;
         static JsonSerializer serializer;
-        static JsonSerializer serializer2;
+        static byte[] data;
     }
 }