+++ src/Elders.Cronus.Serialization.NewtonsoftJson.Tests/runtime-type-inheritence-model/When_serializing_nested_type_undefinded_object_inheritance.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using Machine.Specifications;
 
 namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests
@@ -8,20 +7,16 @@ namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests
     [Subject(typeof(JsonSerializer))]
     public class When_serializing_nested_type_undefinded_object_inheritance
     {
-
         Establish context = () =>
         {
             ser = new NestedTypeWithUndefinedObjectInheritance() { Int = 5, Date = DateTime.UtcNow.AddDays(1), String = "a", Nested = new UndefinedObjectInheritance() { Int = 4, Date = DateTime.UtcNow.AddDays(2), String = "b" } };
             var contracts = new List<Type>();
             contracts.AddRange(typeof(NestedType).Assembly.GetExportedTypes());
             serializer = new JsonSerializer(contracts);
-            serializer2 = new JsonSerializer(contracts);
-            serStream = new MemoryStream();
-            serializer.Serialize(serStream, ser);
-            serStream.Position = 0;
+            data = serializer.SerializeToBytes(ser);
         };
-        Because of_deserialization = () => { deser = (NestedTypeWithUndefinedObjectInheritance)serializer2.Deserialize(serStream); };
 
+        Because of_deserialization = () => deser = serializer.DeserializeFromBytes<NestedTypeWithUndefinedObjectInheritance>(data);
 
         It should_not_be_null = () => deser.ShouldNotBeNull();
         It should_have_the_same_int = () => deser.Int.ShouldEqual(ser.Int);
@@ -32,12 +27,9 @@ namespace Elders.Cronus.Serialization.NewtonsoftJson.Tests
         It nested_object_should_not_be_null = () => deser.Nested.ShouldNotBeNull();
         It nested_object_should_be_of_the_right_type = () => deser.Nested.ShouldBeOfExactType(typeof(UndefinedObjectInheritance));
 
-
-
         static NestedTypeWithUndefinedObjectInheritance ser;
         static NestedTypeWithUndefinedObjectInheritance deser;
-        static Stream serStream;
-        static JsonSerializer serializer;
-        static JsonSerializer serializer2;
+        static JsonSerializer serializer; static JsonSerializer serializer2;
+        static byte[] data;
     }
 }